fix(payment): resolve logging ambiguity and docker networking conflicts
- Corrected SLF4J logger signature in JwtAuthenticationFilter to handle Throwables properly
- Updated application-dev.yaml with refined logging levels (com.itc.funkart: DEBUG)
- Fixed Kafka networking in docker-compose using dual-listener setup (9092 internal/9094 external)
- Added Postgres health checks and internal DNS mapping for service connectivity

+++ backend/payment-service/src/main/java/com/itc/funkart/payment/dto/request/CreatePaymentIntentRequest.java
@@ -3,16 +3,24 @@ package com.itc.funkart.payment.dto.request;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
 
-import java.math.BigDecimal;
-
+/**
+ * Request DTO for initiating a new Stripe PaymentIntent.
+ * <p>
+ * DESIGN NOTES:
+ * 1. <b>Cents-Only:</b> The 'amount' field is a Long to match Stripe's requirement
+ * for the smallest currency unit. This prevents floating-point rounding errors.
+ * 2. <b>Validation:</b> Uses @Positive to ensure we never attempt to charge
+ * a zero or negative balance.
+ * </p>
+ */
 public record CreatePaymentIntentRequest(
         @NotNull(message = "Order ID is required")
         Long orderId,
 
         @NotNull(message = "Amount is required")
         @Positive(message = "Amount must be positive")
-        BigDecimal amount,
+        Long amount, // Represented in cents (e.g., 100 for $1.00)
 
         @NotNull(message = "Currency is required")
-        String currency
+        String currency // ISO code (e.g., "usd")
 ) {}
